feat: validação no delete

+++ src/main/java/com/ada/dynamo/service/QuadroService.java
@@ -2,15 +2,15 @@ package com.ada.dynamo.service;
 
 import com.ada.dynamo.dto.request.QuadroRequest;
 import com.ada.dynamo.dto.response.QuadroResponse;
+import com.ada.dynamo.exception.ItemComAssociassaoException;
 import com.ada.dynamo.exception.ItemNaoEncontradoException;
 import com.ada.dynamo.model.Quadro;
 import com.ada.dynamo.repository.QuadroRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
@@ -19,10 +19,13 @@ public class QuadroService implements ServiceContract<QuadroRequest, QuadroRespo
     private final QuadroRepository repository;
 
     public QuadroResponse findById(String id) {
-        var quadro = repository.findById(id.toString())
-            .orElseThrow(() -> new ItemNaoEncontradoException(String.format("Quadro com id %s não encontrado", id)));
-
+        Quadro quadro = findModelById(id);
         return mapToResponse(quadro);
+    }
+
+    public Quadro findModelById(String id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ItemNaoEncontradoException(String.format("Quadro com id %s não encontrado", id)));
     }
 
     public List<QuadroResponse> findAll() {
@@ -44,28 +47,24 @@ public class QuadroService implements ServiceContract<QuadroRequest, QuadroRespo
     }
 
     public void delete(String id) {
-        var quadro = repository.findById(id).orElseThrow(
-            () -> new ItemNaoEncontradoException(String.format("Quadro com id %s não encontrado para deleção", id))
-        );
+        List<Quadro> partial = repository.findByPartialId(id);
+        if (!partial.isEmpty()) {
+            throw new ItemComAssociassaoException(String.format("O id %s está associado a outros itens e não pode ser deletado", id));
+        }
+
+        Quadro quadro = findModelById(id);
         repository.delete(quadro);
     }
 
     private QuadroResponse mapToResponse(Quadro quadro) {
         var quadroResponse = new QuadroResponse();
-
-        quadroResponse.setId(quadro.getId());
-        quadroResponse.setNome(quadro.getNome());
-        quadroResponse.setTipo(quadro.getTipo());
-
+        BeanUtils.copyProperties(quadro, quadroResponse);
         return quadroResponse;
     }
     
     private Quadro mapToModel(QuadroRequest request) {
         var quadro = new Quadro();
-
-        quadro.setTipo(request.getTipo());
-        quadro.setNome(request.getNome());
-
+        BeanUtils.copyProperties(request, quadro);
         return quadro;
     }
 }
